+++ src/TestSuite/IntegrationTestCase.php
@@ -1,0 +1,38 @@
+<?php
+/**
+ * This file is part of Assets.
+ *
+ * Licensed under The MIT License
+ * For full copyright and license information, please see the LICENSE.txt
+ * Redistributions of files must retain the above copyright notice.
+ *
+ * @copyright   Copyright (c) Mirko Pagliai
+ * @link        https://github.com/mirko-pagliai/assets
+ * @license     https://opensource.org/licenses/mit-license.php MIT License
+ * @since       1.1.8
+ */
+namespace Assets\TestSuite;
+
+use Cake\Core\Configure;
+use Cake\TestSuite\IntegrationTestCase as CakeIntegrationTestCase;
+
+/**
+ * IntegrationTestCaseTest class
+ */
+abstract class IntegrationTestCase extends CakeIntegrationTestCase
+{
+    /**
+     * Teardown any static object changes and restore them
+     * @return void
+     */
+    public function tearDown()
+    {
+        parent::tearDown();
+
+        //Deletes all assets
+        foreach (glob(Configure::read(ASSETS . '.target') . DS . '*') as $file) {
+            //@codingStandardsIgnoreLine
+            @unlink($file);
+        }
+    }
+}