+++ src/Utility/AssetsCreator.php
@@ -16,8 +16,9 @@ namespace Assets\Utility;
 use Cake\Core\Configure;
 use Cake\Core\Plugin;
 use Cake\Filesystem\File;
-use Cake\Network\Exception\InternalErrorException;
+use InvalidArgumentException;
 use MatthiasMullie\Minify;
+use RuntimeException;
 
 /**
  * An utility to create assets
@@ -47,7 +48,7 @@ class AssetsCreator
      * @param string|array $paths String or array of css files
      * @param string $type Extension (`css` or `js`)
      * @return $this
-     * @throws InternalErrorException
+     * @throws InvalidArgumentException
      * @uses resolveAssetPath()
      * @uses resolveFilePaths()
      * @uses $asset
@@ -57,7 +58,7 @@ class AssetsCreator
     public function __construct($paths, $type)
     {
         if (!in_array($type, ['css', 'js'])) {
-            throw new InternalErrorException(__d('assets', 'Asset type `{0}` not supported', $type));
+            throw new InvalidArgumentException(__d('assets', 'Asset type `{0}` not supported', $type));
         }
 
         //Note: `resolveFilePaths()` method needs `$type` property;
@@ -88,7 +89,7 @@ class AssetsCreator
      * Internal method to resolve partial file paths and return full paths
      * @param string|array $paths Partial file paths
      * @return array Full file paths
-     * @throws InternalErrorException
+     * @throws RuntimeException
      * @use $type
      */
     protected function resolveFilePaths($paths)
@@ -105,15 +106,14 @@ class AssetsCreator
             }
 
             $path = substr($path, 0, 1) === '/' ? substr($path, 1) : $this->type . DS . $path;
+            $path = DS === '/' ? $path : $path = str_replace('/', DS, $path);
             $path = empty($plugin) ? WWW_ROOT . $path : Plugin::path($plugin) . 'webroot' . DS . $path;
 
             //Appends the file extension, if not already present
-            if (pathinfo($path, PATHINFO_EXTENSION) !== $this->type) {
-                $path = sprintf('%s.%s', $path, $this->type);
-            }
+            $path = pathinfo($path, PATHINFO_EXTENSION) == $this->type ? $path : sprintf('%s.%s', $path, $this->type);
 
-            if (!file_exists($path)) {
-                throw new InternalErrorException(__d('assets', 'File `{0}` doesn\'t exist', str_replace(APP, null, $path)));
+            if (!is_readable($path)) {
+                throw new RuntimeException(__d('assets', 'File `{0}` doesn\'t exist', rtr($path)));
             }
 
             return $path;
@@ -123,7 +123,7 @@ class AssetsCreator
     /**
      * Creates the asset
      * @return string
-     * @throws InternalErrorException
+     * @throws RuntimeException
      * @uses filename()
      * @uses path()
      * @uses $paths
@@ -131,9 +131,7 @@ class AssetsCreator
      */
     public function create()
     {
-        $asset = $this->path();
-
-        if (!is_readable($asset)) {
+        if (!is_readable($this->path())) {
             switch ($this->type) {
                 case 'css':
                     $minifier = new Minify\CSS();
@@ -146,8 +144,8 @@ class AssetsCreator
             array_map([$minifier, 'add'], $this->paths);
 
             //Writes the file
-            if (!(new File($asset, false, 0755))->write($minifier->minify())) {
-                throw new InternalErrorException(__d('assets', 'Failed to create file {0}', str_replace(APP, null, $asset)));
+            if (!(new File($this->path(), false, 0755))->write($minifier->minify())) {
+                throw new RuntimeException(__d('assets', 'Failed to create file {0}', rtr($this->path())));
             }
         }
 
@@ -157,10 +155,11 @@ class AssetsCreator
     /**
      * Returns the asset filename
      * @return string Asset filename
+     * @uses path()
      */
     public function filename()
     {
-        return pathinfo($this->asset, PATHINFO_FILENAME);
+        return pathinfo($this->path(), PATHINFO_FILENAME);
     }
 
     /**